requirements.txt
py7zr
requests
requests-cache
python-dotenv-vault
Flask==2.3.2
Flask-Classful
Flask-Compress
humanfriendly
jinja2-env
jinja2-humanize-extension
jinja-markdown
Pympler
better-pyenv
graphql-core
orjson
rich
dotted-dict
heroicons
jinja2-simple-tags
cssmin
htmlmin
jsmin
Brotli
wcmatch
jmespath
jsonurl-py
arrow
numpy
python-liquid
minify_html
Werkzeug==2.3.6
tzdata
tzlocal